daemon: split out sections support to api_sections.go
apparently this had no tests so far, added a TODO for now
